Refactor pydicom RLE implementation
