+++ pydicom/encoders/pylibjpeg.py
@@ -1,3 +1,4 @@
+# Copyright 2008-2021 pydicom authors. See LICENSE file for details.
 """Interface for *Pixel Data* encoding, not intended to be used directly."""
 
 from pydicom.uid import RLELossless
